+++ inst/templates/ad_adpc.R
@@ -72,6 +72,7 @@ pc_dates <- pc %>%
   ) %>%
   derive_var_nfrlt(
     new_var = NFRLT,
+    out_unit = "hours",
     tpt_var = PCTPT,
     visit_day = VISITDY,
     treatment_duration = 0,
@@ -108,6 +109,7 @@ ex_dates <- ex %>%
   ) %>%
   derive_var_nfrlt(
     new_var = NFRLT,
+    out_unit = "hours",
     visit_day = VISITDY
   ) %>%
   # Set missing end dates to start date